First draft of SOPS parser integration
Downside is that the core logic only assumes there's ansible-vault
integration, so the current implementation will assume that sops is in
the PATH. If it's not, I expect a pretty ugly failure.

Includes rather robust error handling. For a vars file to be interpreted
as a SOPS-encrypted, all the following must be true:

  * must be YAML format (current limitation)
  * must have top-level "sops" key
  * top-level "sops" key must have subkeys:
    * mac
    * version
    * lastmodified

Parser logic will raise informative error messages, e.g. if sops file
parsing is attempted but sops is not present in PATH.